\name{AirBcn}
\alias{AirBcn}
\docType{data}
\title{
  Monthly Airline Passenger Numbers 1990-2009 of Barcelona
}
\description{
   Monthly totals of international airline passengers of Barcelona, 1990 to 2009.
}
\usage{AirBcn}
\format{
  A monthly time series, in thousands.
}
\source{
  Ministerio de Fomento (www.fomento.es)
}

\examples{
	data(AirBcn)
	plot(AirBcn)
}
\keyword{datasets}