\name{TSTutorial-package}
\alias{TSTutorial-package}
\alias{[,cami,character,missing,missing-method}
\alias{[,datos,character,missing,missing-method}
\alias{[,lineal,character,missing,missing-method}
\alias{[,menu,character,missing,missing-method}
\alias{[,menuitem,character,missing,missing-method}
\alias{[,modelo,character,missing,missing-method}
\alias{[,report,character,missing,missing-method}
\alias{[,serie,character,missing,missing-method}
\alias{[,sessio.ts,character,missing,missing-method}
\alias{[,tex,character,missing,missing-method}
\alias{[<-,cami,character,missing-method}
\alias{[<-,datos,character,missing-method}
\alias{[<-,lineal,character,missing-method}
\alias{[<-,menu,character,missing-method}
\alias{[<-,menuitem,character,missing-method}
\alias{[<-,modelo,character,missing-method}
\alias{[<-,report,character,missing-method}
\alias{[<-,serie,character,missing-method}
\alias{[<-,sessio.ts,character,missing-method}
\alias{[<-,tex,character,missing-method}

\docType{package}
\title{
Time Series Tutorial
}
\description{
\code{TSTutorial} packages is an implementation for learning fitting and getting predictions of Time Series with the methology Box-Jenkins. This consist on an interactive laboratory showing differents menus which contain all the steps of the methodology. Furthemore, these menus, and its options, have some explanatory texts and suggestions (based in created tests) to facilitate the learning process.
}
\details{
\tabular{ll}{
Package: \tab TSTutorial\cr
Type: \tab Package\cr
Version: \tab 1.2.1\cr
Date: \tab 2012-11-22\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
Depends: \tab methods, grDevices, graphics, stats, snowfall, MASS\cr
}
}
\section{Overview}{
	The use of this package is very simple, to start the tutorial only is required a time series in the method \code{TSTutorial}.
	
	Moreover, the package includes three time series as data examples to use them with the function \code{TSTutorial}.
	
	
}


\author{
	Author and maintainer: Alberto Lopez Moreno <bertolomo@gmail.com>\cr
}


\seealso{
Overview: \code{\link{TSTutorial-package}}.\cr
Methods	: \code{\link{TSTutorial}}.\cr
}

\keyword{ package }