\name{Turismes}
\alias{Turismes}
\docType{data}
\title{
  Monthly Made Vehicles in Spain 1990-2008.
}
\description{
   Monthly made vehicles in Spain, Genuary 1990 to August 2008.
}
\usage{Turismes}
\format{
    A monthly time series, in thousands.
}
\source{
  Instituto Nacional de Estadistica (www.ine.es).
}

\examples{
	data(Turismes)
	plot(Turismes)
}
\keyword{datasets}