\name{Victimes}
\alias{Victimes}
\docType{data}
\title{
  Monthly Traffic Deads Number 1993-2008 of Spain.
}
\description{
  Monthly totals of traffic deads in road and urban accidents of Spain, 1993 to 2008.
}
\usage{Victimes}
\format{
  A monthly time series, in thousands.
}
\source{
  Direccion General de Trafico (www.dgt.es)
}

\examples{
	data(Victimes)
	plot(Victimes)
}
\keyword{datasets}